requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
aio-pika>=9.0.0
aiokafka>=0.8.0
redis>=5.0.0
asyncpg>=0.28.0
sqlalchemy>=2.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
jsonschema>=4.18.0
uuid>=1.30
requests>=2.31.0
beautifulsoup4>=4.12.0
python-whois>=0.8.0
dnspython>=2.4.0
snscrape
python-jose[cryptography]
passlib[bcrypt]
python-multipart
